childAllowanceCalculator
In this program, I used a simple if statement to assign how much child allowance a family gets. the two variables are how much the family earns and how many children in the family. Depending on these variables, the program answers how much a family is due in tax allowances.