replacingVowels P4_3d
In this example I showcase my ability to use a while loop combined with string manipulation and an if statement to take any word and replace the vowels with _. Boolean logic is also used here. 

For example, it you type "Hello" it will spit out "H_ll_".  

Notice on lines 14 and 15 that I account for both upper and lowercase instances.